+++ src/Simulation/Simulator.java
@@ -14,38 +14,27 @@ import java.util.Random;
 public class Simulator {
 	private static final String AD_HOC = "1";
 	private static final String PASS = "2";
-	private CarQueue entranceCarQueue;
-    private CarQueue entrancePassQueue;
-    private CarQueue paymentCarQueue;
-    private CarQueue exitCarQueue;
-    private SimulatorView simulatorView;
     private Daytime startDayTime;
     private Daytime endDayTime;
     private DaytimeSimulator daytimeSimulator;
+    private GarageSimulator garageSimulator;
     private int tickPause = 100;
 
     int weekDayArrivals= 100; // average number of arriving cars per hour
     int weekendArrivals = 200; // average number of arriving cars per hour
     int weekDayPassArrivals= 50; // average number of arriving cars per hour
     int weekendPassArrivals = 5; // average number of arriving cars per hour
-    int enterSpeed = 3; // number of cars that can enter per minute
-    int paymentSpeed = 7; // number of cars that can pay per minute
-    int exitSpeed = 5; // number of cars that can leave per minute
 
     /**
      * The Simulator constructor
      * @author Hanzehogeschool of Applied Sciences
      */
     public Simulator() {
-        this.entranceCarQueue = new CarQueue();
-        this.entrancePassQueue = new CarQueue();
-        this.paymentCarQueue = new CarQueue();
-        this.exitCarQueue = new CarQueue();
-        this.simulatorView = new SimulatorView(3, 6, 30);
         this.startDayTime = new Daytime();
         this.endDayTime = new Daytime();
         this.endDayTime.hours = 1;
         this.daytimeSimulator = new DaytimeSimulator(this.startDayTime);
+        this.garageSimulator = new GarageSimulator();
     }
 
     /**
@@ -68,8 +57,10 @@ public class Simulator {
      */
     private void tick() {
         this.daytimeSimulator.tick();
-        this.handleExit();
-        this.updateViews();
+
+        this.carsArriving();
+
+        this.garageSimulator.tick();
 
     	// Pause.
         try {
@@ -77,39 +68,6 @@ public class Simulator {
         } catch (InterruptedException e) {
             e.printStackTrace();
         }
-
-        this.handleEntrance();
-    }
-
-    /**
-     * Update entrance queues
-     * @author Hanzehogeschool of Applied Sciences
-     */
-    private void handleEntrance() {
-        this.carsArriving();
-        this.carsEntering(this.entrancePassQueue);
-        this.carsEntering(this.entranceCarQueue);
-    }
-
-    /**
-     * Update exit queues
-     * @author Hanzehogeschool of Applied Sciences
-     */
-    private void handleExit() {
-        this.carsReadyToLeave();
-        this.carsPaying();
-        this.carsLeaving();
-    }
-
-    /**
-     * Update SimultorView
-     * @author Hanzehogeschool of Applied Sciences
-     */
-    private void updateViews() {
-        this.simulatorView.tick();
-
-        // Update the car park view.
-        this.simulatorView.updateView();
     }
 
     /**
@@ -117,82 +75,13 @@ public class Simulator {
      * @author Hanzehogeschool of Applied Sciences
      */
     private void carsArriving() {
-    	int numberOfCars = this.getNumberOfCars(this.weekDayArrivals, this.weekendArrivals);
+        int numberOfCars = this.getNumberOfCars(this.weekDayArrivals, this.weekendArrivals);
 
         this.addArrivingCars(numberOfCars, AD_HOC);
 
-    	numberOfCars = this.getNumberOfCars(this.weekDayPassArrivals, this.weekendPassArrivals);
+        numberOfCars = this.getNumberOfCars(this.weekDayPassArrivals, this.weekendPassArrivals);
 
         this.addArrivingCars(numberOfCars, PASS);
-    }
-
-    /**
-     * Update entering cars
-     * @author Hanzehogeschool of Applied Sciences
-     * @param queue
-     */
-    private void carsEntering(CarQueue queue) {
-        int i = 0;
-
-        // Remove car from the front of the queue and assign to a parking space.
-    	while (queue.carsInQueue() > 0 && this.simulatorView.getNumberOfOpenSpots() > 0 && i < this.enterSpeed) {
-            Car car = queue.removeCar();
-            Location freeLocation = this.simulatorView.getFirstFreeLocation();
-
-            this.simulatorView.setCarAt(freeLocation, car);
-            ++i;
-        }
-    }
-
-    /**
-     * Update cars ready to leave
-     * @author Hanzehogeschool of Applied Sciences
-     */
-    private void carsReadyToLeave() {
-        // Add leaving cars to the payment queue.
-        Car car = this.simulatorView.getFirstLeavingCar();
-
-        while (car != null) {
-        	if (car.getHasToPay()) {
-	            car.setIsPaying(true);
-                this.paymentCarQueue.addCar(car);
-        	} else {
-                this.carLeavesSpot(car);
-        	}
-
-            car = this.simulatorView.getFirstLeavingCar();
-        }
-    }
-
-    /**
-     * Update paying cars
-     * @author Hanzehogeschool of Applied Sciences
-     */
-    private void carsPaying() {
-        // Let cars pay.
-    	int i = 0;
-
-    	while (this.paymentCarQueue.carsInQueue() > 0 && i < this.paymentSpeed) {
-            Car car = this.paymentCarQueue.removeCar();
-
-            // TODO Handle payment.
-            this.carLeavesSpot(car);
-            ++i;
-    	}
-    }
-
-    /**
-     * Update leaving cars
-     * @author Hanzehogeschool of Applied Sciences
-     */
-    private void carsLeaving(){
-        // Let cars leave.
-    	int i = 0;
-
-    	while (this.exitCarQueue.carsInQueue() > 0 && i < this.exitSpeed) {
-            this.exitCarQueue.removeCar();
-            ++i;
-    	}	
     }
 
     /**
@@ -231,28 +120,18 @@ public class Simulator {
      */
     private void addArrivingCars(int numberOfCars, String type) {
         // Add the cars to the back of the queue.
-    	switch(type) {
-    	case AD_HOC: 
-            for (int i = 0; i < numberOfCars; ++i) {
-                this.entranceCarQueue.addCar(new AdHocCar());
-            }
-            break;
+        switch(type) {
+            case AD_HOC:
+                for (int i = 0; i < numberOfCars; ++i) {
+                    this.garageSimulator.addArrivingAdHocCar();
+                }
+                break;
 
-    	case PASS:
-            for (int i = 0; i < numberOfCars; ++i) {
-                this.entrancePassQueue.addCar(new ParkingPassCar());
-            }
-            break;	            
-    	}
-    }
-
-    /**
-     * Make the car leave their spot
-     * @author Hanzehogeschool of Applied Sciences
-     * @param car the car to leave it's spot
-     */
-    private void carLeavesSpot(Car car) {
-        this.simulatorView.removeCarAt(car.getLocation());
-        this.exitCarQueue.addCar(car);
+            case PASS:
+                for (int i = 0; i < numberOfCars; ++i) {
+                    this.garageSimulator.addArrivingPassCar();
+                }
+                break;
+        }
     }
 }
